Use imported data
