+++ src/data/parking.js
@@ -1,0 +1,23 @@
+const options = {
+  free: [
+    {
+      name: 'Rapee Sagarik Building',
+      location: 'Behind Rapee Sagarik Building',
+      capacity: 40,
+    },
+  ],
+  paid: [
+    {
+      name: 'Ngam Wong Wan 1 Parking Building',
+      location: 'Ngam Wong Wan 1 entrance',
+      capacity: 289,
+    },
+
+    {
+      name: 'Ngam Wong Wan 2 Parking Building',
+      location: 'Ngam Wong Wan 2 entrance',
+      capacity: 192,
+    },
+  ],
+}
+export default options
